* Use normal php travis mechanism instead of rake, so when errors normal script execution status code will be returned. DoctrineBundle repo was moved, fixed vendor/vendors.php script
* Skipped test if some GD and FreeType function do not exist

* Added install command to vendor/vendors

* Skip GD tests when gd with freetype is not installed
