+++ Tests/Form/Type/JQuery/FileTypeTest.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Genemu\Bundle\FormBundle\Tests\Form\Type;
+namespace Genemu\Bundle\FormBundle\Tests\Form\Type\JQuery;
 
 use Symfony\Component\HttpFoundation\File\File;
